(perform some Refactoring for clarity)
