+++ src/m3/test/self-definition.ts
