+++ src/utils/sorting.ts
@@ -1,0 +1,16 @@
+/**
+ * Various functional utilities for sorting things.
+ */
+
+
+export type Comparer<T> = (l: T, r: T) => number
+
+const stringCompare: Comparer<string> = (l, r): number =>
+    l === r ? 0 : (l > r ? 1 : -1)
+
+export const stringyCompare = <T>(keyFunc: (t: T) => string): Comparer<T> =>
+    (l: T, r: T) => stringCompare(keyFunc(l), keyFunc(r))
+
+export const sortByStringKey = <T>(ts: T[], keyFunc: (t: T) => string) =>
+    [...ts].sort(stringyCompare(keyFunc))
+
