+++ watch-test-m3.sh
@@ -1,2 +1,2 @@
-mkdir -p plantUML
+mkdir -p diagrams
 deno test --watch --allow-write src/m3/test/test-*.ts
